fix(engine-loader): avoid stale `asset-loader`
For some reason the injection of the `asset-loader` becomes "stale" in
production builds. It seems like, the service is looked up first, then
re-registered by a third-party and the injection is not updated.

This causes the manifest to be empty and the loading to always fail.

I have no idea what the root cause is, but this trivial change is safe
and makes sure that we always look up the "feshest" version of the
service.